+++ mdsf/src/tools/grafbase_lint.rs
@@ -1,0 +1,26 @@
+///
+/// THIS FILE IS GENERATED USING CODE - DO NOT EDIT MANUALLY
+///
+use crate::runners::CommandType;
+
+#[inline]
+pub fn set_args(
+    mut cmd: std::process::Command,
+    file_path: &std::path::Path,
+) -> std::process::Command {
+    cmd.arg("lint");
+    cmd.arg(file_path);
+    cmd
+}
+
+pub const COMMANDS: [CommandType; 7] = [
+    CommandType::NodeModules("grafbase"),
+    CommandType::Direct("grafbase"),
+    CommandType::Npm("grafbase"),
+    CommandType::Pnpm("grafbase"),
+    CommandType::Bun("grafbase"),
+    CommandType::Deno("grafbase"),
+    CommandType::Yarn("grafbase"),
+];
+
+pub const IS_STDIN: bool = false;